* SideMenuToolbar --> MainView * Drawer and Toolbar components extracted from MainView
